Enable secondary_user_on_secondary_display for CtsLiblogTestCases
--secondary_user_on_secondary_display is for background users that have
UI access on assigned displays, a.k.a. visible background users, on
devices that have config_multiuserVisibleBackgroundUsers enabled.

The main use case is Automotive's multi-display where passengers
(modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) who interacts with driver's display.

Bug: 354264010

Test: atest --user-type secondary_user_on_secondary_display \
          CtsLiblogTestCases
Test: atest CtsLiblogTestCases

Change-Id: I6686e7cdd89b36ab9d5c32ae3fa063cb9cee490f