Add log ID to proto output.
This allows for tooling that consumes the proto output to make decisions
based on log source without having to run one copy of logcat per log
source. For example, most of the kernel log tags are broken, so we may
want to apply some better heuristics for determining tag, or throw out
the idea of a tag altogether.

Test: Inspect logcat proto output

Change-Id: Ibca34fac712defd32944cfeae0efc44cac68be1b